service/requirements.txt
# Core framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0

# MCP server (SSE transport)
mcp[cli]>=1.3.0

# Configuration
python-dotenv>=1.0.0

# HTTP client (for proxy and calling sub-services)
httpx>=0.27.0

# Docker SDK (for managing sub-containers)
docker>=7.0.0

# Database
sqlalchemy>=2.0.31

# Memory system (mem0 dependencies)
qdrant-client>=1.9.1
openai>=1.40.0
ollama>=0.6.1

# Graph memory
langchain-neo4j>=0.4.0
neo4j>=5.23.1
rank-bm25>=0.2.2

# Memory utilities
tenacity==9.1.2
fastapi-pagination>=0.12.0
python-multipart>=0.0.5

# mem0 core deps
posthog>=3.5.0
pytz>=2024.1
protobuf>=5.29.0,<6.0.0
